+++ core-ui/src/test/java/com/agileburo/anytype/core_ui/BlockViewDiffUtilTest.kt
@@ -407,4 +407,41 @@ class BlockViewDiffUtilTest {
             actual = payload
         )
     }
+
+    @Test
+    fun `should detect focus change for title block`() {
+
+        val index = 0
+
+        val id = MockDataFactory.randomUuid()
+
+        val text = MockDataFactory.randomString()
+
+        val oldBlock = BlockView.Title(
+            id = id,
+            text = text,
+            focused = false
+        )
+
+        val newBlock: BlockView = oldBlock.copy(
+            focused = true
+        )
+
+        val old = listOf(oldBlock)
+
+        val new = listOf(newBlock)
+
+        val diff = BlockViewDiffUtil(old = old, new = new)
+
+        val payload = diff.getChangePayload(index, index)
+
+        val expected = Payload(
+            changes = listOf(BlockViewDiffUtil.FOCUS_CHANGED)
+        )
+
+        assertEquals(
+            expected = expected,
+            actual = payload
+        )
+    }
 }